DAOs modified

+++ src/sql/create-schema.sql
@@ -22,7 +22,7 @@ DROP SEQUENCE IF EXISTS blog_id_seq;
 CREATE SEQUENCE blog_id_seq;
 
 CREATE TABLE blog (
-	id_blog INTEGER DEFAULT NEXTVAL('blog_id_seq'),
+	id_blog BIGINT DEFAULT NEXTVAL('blog_id_seq'),
 	titulo_blog VARCHAR(20),
 	fecha_creacion_blog TIMESTAMP,
 	usuario_blog VARCHAR(20),
@@ -36,12 +36,12 @@ DROP SEQUENCE IF EXISTS articulo_id_seq;
 CREATE SEQUENCE articulo_id_seq;
 
 CREATE TABLE articulo (
-	id_articulo INTEGER DEFAULT NEXTVAL('articulo_id_seq'),
+	id_articulo BIGINT DEFAULT NEXTVAL('articulo_id_seq'),
 	titulo_articulo VARCHAR(20),
 	fecha_publicacion_articulo TIMESTAMP,
 	texto_articulo VARCHAR(100),
 	me_gusta_articulo BOOLEAN,
-	blog_articulo INTEGER,
+	blog_articulo BIGINT,
 	CONSTRAINT articulo_pk PRIMARY KEY (id_articulo),
 	CONSTRAINT articulo_fk FOREIGN KEY (blog_articulo) REFERENCES blog(id_blog)
 );
@@ -51,13 +51,13 @@ DROP SEQUENCE IF EXISTS enlace_id_seq;
 CREATE SEQUENCE enlace_id_seq;
 
 CREATE TABLE enlace (
-	id_enlace INTEGER DEFAULT NEXTVAL('enlace_id_seq'),
+	id_enlace BIGINT DEFAULT NEXTVAL('enlace_id_seq'),
 	titulo_enlace VARCHAR(20),
 	fecha_publicacion_enlace TIMESTAMP,
 	url_enlace VARCHAR(20),
 	tipo_contenido_enlace VARCHAR(20),
 	me_gusta_enlace BOOLEAN,
-	blog_enlace INTEGER,
+	blog_enlace BIGINT,
 	CONSTRAINT enlace_pk PRIMARY KEY (id_enlace),
 	CONSTRAINT enlace_fk FOREIGN KEY (blog_enlace) REFERENCES blog(id_blog)
 );
